move `popupAutomaticMembershipAlert` funtion to groups-new controller
