+++ app/assets/javascripts/discourse/components/group-manage-save-button.js
@@ -1,7 +1,7 @@
 import discourseComputed from "discourse-common/utils/decorators";
 import Component from "@ember/component";
 import { popupAjaxError } from "discourse/lib/ajax-error";
-import { popupAutomaticMembershipAlert } from "discourse/lib/utilities";
+import { popupAutomaticMembershipAlert } from "discourse/controllers/groups-new";
 
 export default Component.extend({
   saving: null,
